Now a semifunctional thermostat
Added basic, hardcoded thermostat functionality. Every ten minutes, it
wakes up and checks the temperature. If it's too cold, it turns on the
heat until it's the correct temperature again. Also uses the Particle
Cloud for basic logging, but does NOT manage to send events when it
doesn't turn the heat on; it goes to sleep too quickly.